requirements.txt
librosa==0.9.2
matplotlib==3.5.1
munch==2.5.0
ninja==1.10.2.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
scikit-image==0.19.2
tqdm==4.64.0
timm==0.6.13
face-alignment==1.3.5
paddlepaddle==2.4.2